problem evaulator implemented; application context - Main bean removed

+++ src/main/java/pl/edu/agh/muvto/model/MuvtoEdge.java
@@ -27,7 +27,7 @@ public class MuvtoEdge extends EntityWithId {
     }
 
     public double getWeight() {
-        return getFill()/getCapacity();
+        return getFill()/(double)getCapacity();
     }
     
     public double getAttractiveness() {
